requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pillow
python-multipart
